# Conflicts:
#	Source/MegaBonkLike/Player/MBLPlayerController.h

+++ Source/MegaBonkLike/Gimmick/Objects/SpawnObjects/MBLExpObject.h
@@ -15,5 +15,5 @@ public:
 
 private:
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "ExpObject", meta = (AllowPrivateAccess = "true"))
-	int32 Exp;
+	float Exp;
 };
